Separate hook for swap refocus
